feat: add TodaysEtrianDateDisplay component and integrate it into the EtrianCalendar page
