+++ src/app/(toys)/etrian-calendar/page.mdx
@@ -1,7 +1,10 @@
+import { TodaysEtrianDateDisplay } from "./_components/todays-etrian-date-display";
 import { EtrianCalendar } from "./_components/etrian-calendar";
 import { Version } from "@/components/shared/version";
 
 # 世界樹の暦
+
+<TodaysEtrianDateDisplay />
 
 <EtrianCalendar />
